+++ CHANGELOG.md
@@ -11,7 +11,7 @@ CHANGELOG for FlatCAM beta
 
 - NCC Tool optimization - moved the UI in its own class
 - NCC Tool optimization - optimized the Tool edit method
-
+- NCC Tool - allow no tool at NCC Tool start (the Preferences have no tool)
 
 11.06.2020